+++ sw/airborne/modules/nav/nav_rotorcraft_base.c
@@ -24,7 +24,7 @@
  * Basic navigation functions for Rotorcraft
  */
 
-#include "modules/nav/nav_rover_base.h"
+#include "modules/nav/nav_rotorcraft_base.h"
 
 struct NavBase_t nav_rotorcraft_base;
 
@@ -247,7 +247,6 @@ static void nav_oval(struct EnuCoor_f *wp1, struct EnuCoor_f *wp2, float radius)
 #if PERIODIC_TELEMETRY
 #include "modules/datalink/telemetry.h"
 
-#if ROVER_BASE_SEND_TRAJECTORY
 static void send_segment(struct transport_tx *trans, struct link_device *dev)
 {
   pprz_msg_send_SEGMENT(trans, dev, AC_ID,
@@ -264,7 +263,6 @@ static void send_circle(struct transport_tx *trans, struct link_device *dev)
       &nav_rotorcraft_base.circle.center.y,
       &nav_rotorcraft_base.circle.radius);
 }
-#endif // ROVER_BASE_SEND_TRAJECTORY
 
 static void send_nav_status(struct transport_tx *trans, struct link_device *dev)
 {
@@ -275,13 +273,11 @@ static void send_nav_status(struct transport_tx *trans, struct link_device *dev)
                                       &dist_home, &dist_wp,
                                       &nav_block, &nav_stage,
                                       &nav.mode);
-#if ROVER_BASE_SEND_TRAJECTORY
   if (nav.mode == NAV_HORIZONTAL_MODE_ROUTE) {
     send_segment(trans, dev);
   } else if (nav.mode == NAV_HORIZONTAL_MODE_CIRCLE) {
     send_circle(trans, dev);
   }
-#endif // ROVER_BASE_SEND_TRAJECTORY
 }
 #endif
